JZ: Initial endpoint/view commit + more tests

+++ onadata/libs/utils/csv_import.py
@@ -2,9 +2,10 @@ import unicodecsv as ucsv
 import uuid
 import json
 
-from cStringIO import StringIO
+import cStringIO
 from datetime import datetime
-from onadata.libs.utils.logger_tools import dict2xml, create_instance
+from django.contrib.auth.models import User
+from onadata.libs.utils.logger_tools import dict2xml, safe_create_instance
 from onadata.apps.logger.models import Instance
 
 
@@ -16,7 +17,7 @@ def get_submission_meta_dict(xform, instance_id):
     uuid_arg = 'uuid:{}'.format(uuid.uuid4())
     meta = {'instanceID': uuid_arg}
 
-    if len(xform.instances.filter(uuid=instance_id)) > 0:
+    if xform.instances.filter(uuid=instance_id).count() > 0:
         uuid_arg = 'uuid:{}'.format(uuid.uuid4())
         meta.update({'instanceID': uuid_arg,
                      'deprecatedID': 'uuid:{}'.format(instance_id)})
@@ -37,19 +38,22 @@ def dict2xmlsubmission(submission_dict, xform, instance_id, submission_date):
 def submit_csv(username, xform, csv_data):
 
     if isinstance(csv_data, (str, unicode)):
-        csv_data = StringIO(csv_data)
-    elif not isinstance(csv_data, file):
+        csv_data = cStringIO.StringIO(csv_data)
+    elif csv_data.read is None:
         raise TypeError(u'Invalid param type for `csv_data`. '
-                        'Expected file, String or Unicode but '
+                        'Expected file or String '
                         'got {} instead.'.format(type(csv_data).__name__))
 
     csv_reader = ucsv.DictReader(csv_data)
     rollback_uuids = []
     submission_time = datetime.utcnow().isoformat()
     ona_uuid = {'formhub': {'uuid': xform.uuid}}
+    error = None
+    count = 0
     for row in csv_reader:
         # fetch submission uuid before purging row metadata
         row_uuid = row.get('_uuid')
+        submitted_by = row.get('_submitted_by')
         submission_date = row.get('_submission_time', submission_time)
 
         for key in row.keys():  # seems faster than a comprehension
@@ -72,15 +76,22 @@ def submit_csv(username, xform, csv_data):
         row_uuid = row.get('meta').get('instanceID')
         rollback_uuids.append(row_uuid.replace('uuid:', ''))
 
-        xml_file = StringIO(dict2xmlsubmission(row, xform, row_uuid,
-                                               submission_date))
+        xml_file = cStringIO.StringIO(dict2xmlsubmission(row, xform, row_uuid,
+                                      submission_date))
 
-        try:
-            create_instance(username, xml_file, [],
-                            u'submitted_via_web', xform.uuid)
-        except:
-            # there has to be a more elegant way to roll back
-            # the following is a stop-gap
+        error, instance = safe_create_instance(username, xml_file, [],
+                                               xform.uuid, None)
+
+        if error:
             Instance.objects.filter(uuid__in=rollback_uuids,
                                     xform=xform).delete()
-            raise
+            return error
+        else:
+            count += 1
+            users = User.objects.filter(
+                username=submitted_by) if submitted_by else []
+            if users:
+                instance.user = users[0]
+                instance.save()
+
+    return count
